requirements.txt
anthropic
openai
python-dotenv
colorama